Move `FadeInWidget`
